requirements.txt
flask==3.0.3
google-generativeai==0.8.6
python-dotenv==1.0.1
flask-limiter==3.5.0
requests==2.31.0
gunicorn==21.2.0